llcppsymg:libs,cflags parse
