+++ chore/_xtool/llcppsymg/_cmptest/config_test/llgo.expect
@@ -18,6 +18,79 @@ Cplusplus: false
 === Test case: Invalid JSON ===
 Error: failed to parse config
 
+=== Test ParseLibs ===
+Test case: Lua library
+Input: -L/opt/homebrew/lib -llua -lm
+Paths: [/opt/homebrew/lib]
+Names: [lua m]
+Test case: SQLite library
+Input: -L/opt/homebrew/opt/sqlite/lib -lsqlite3
+Paths: [/opt/homebrew/opt/sqlite/lib]
+Names: [sqlite3]
+Test case: INIReader library
+Input: -L/opt/homebrew/Cellar/inih/58/lib -lINIReader
+Paths: [/opt/homebrew/Cellar/inih/58/lib]
+Names: [INIReader]
+Test case: Multiple library paths
+Input: -L/opt/homebrew/lib -L/usr/lib -llua
+Paths: [/opt/homebrew/lib /usr/lib]
+Names: [lua]
+Test case: No valid library
+Input: -L/opt/homebrew/lib
+Paths: [/opt/homebrew/lib]
+Names: []
+=== Test GenDylibPaths ===
+Test case: existing dylib
+Path libsymb1 is in the expected paths
+Test case: existing dylibs
+Path libsymb1 is in the expected paths
+Path libsymb2 is in the expected paths
+Test case: existint default paths
+Path libsymb1 is in the expected paths
+Path libsymb3 is in the expected paths
+Test case: existint default paths & not found
+notFounds [math]
+Path libsymb1 is in the expected paths
+Path libsymb3 is in the expected paths
+Test case: no existing dylib
+notFounds [notexist]
+Error: failed to find any libraries
+=== Test ParseCFlags ===
+Test case: Single include path
+Input: -I/usr/include
+Paths: [/usr/include]
+Test case: Multiple include paths
+Input: -I/usr/include -I/opt/homebrew/include
+Paths: [/usr/include /opt/homebrew/include]
+Test case: Include paths mixed with other flags
+Input: -I/usr/include -DDEBUG -I/opt/local/include -Wall
+Paths: [/usr/include /opt/local/include]
+Test case: Empty input
+Input: 
+Paths: []
+=== Test GenHeaderFilePath ===
+Test case: Valid files
+Input files: [test1.h test2.h]
+Output: [test1.h test2.h]
+
+Test case: Mixed existing and non-existing files
+Input files: [test1.h nonexistent.h]
+notFounds [nonexistent.h]
+Output: [test1.h]
+
+Test case: Multiple include paths
+Input files: [test1.h test2.h test3.h]
+Output: [test1.h test2.h test3.h]
+
+Test case: No existing files
+Input files: [nonexistent1.h nonexistent2.h]
+Error: failed to find any header files
+notFounds [nonexistent1.h nonexistent2.h]
+
+Test case: Empty file list
+Input files: []
+Error: failed to find any header files
+
 
 #stderr
