+++ chore/_xtool/llcppsymg/_cmptest/symbol_test/llgo.expect
@@ -1,35 +1,4 @@
 #stdout
-=== Test ParseLibConfig ===
-Test case: Lua library
-Input: -L/opt/homebrew/lib -llua -lm
-Paths: [/opt/homebrew/lib]
-Names: [lua m]
-Test case: SQLite library
-Input: -L/opt/homebrew/opt/sqlite/lib -lsqlite3
-Paths: [/opt/homebrew/opt/sqlite/lib]
-Names: [sqlite3]
-Test case: INIReader library
-Input: -L/opt/homebrew/Cellar/inih/58/lib -lINIReader
-Paths: [/opt/homebrew/Cellar/inih/58/lib]
-Names: [INIReader]
-Test case: Multiple library paths
-Input: -L/opt/homebrew/lib -L/usr/lib -llua
-Paths: [/opt/homebrew/lib /usr/lib]
-Names: [lua]
-Test case: No valid library
-Input: -L/opt/homebrew/lib
-Paths: [/opt/homebrew/lib]
-Names: []
-=== Test GenDylibPaths ===
-Test case: existing dylib
-Path libsymb1 is in the expected paths
-Test case: existing dylibs
-Path libsymb1 is in the expected paths
-Path libsymb2 is in the expected paths
-Test case: existint default paths
-Path libsymb1 is in the expected paths
-Path libsymb3 is in the expected paths
-Test case: no existing dylib
 === Test GetCommonSymbols ===
 
 Test Case: Lua symbols
